requirements.txt
requests>=2.31
ping3>=4.0
python-nmap>=0.7
dnspython>=2.4
rich>=13.0
paramiko>=3.4
